src/flask_structured_api/requirements.txt
# Core Flask
Flask==3.1.0
flask-openapi3==4.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
Werkzeug==3.1.3
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.36
sqlmodel==0.0.22
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth & Security
PyJWT==2.10.0
passlib==1.7.4
flask-cors==4.0.0

# Cache & Queue
redis==5.2.0
celery==5.3.6
flower==2.0.1

# Config & Environment
python-dotenv==1.0.0
email-validator>=2.1.0
pydantic[email]>=2.5.0
pydantic-settings==2.6.1

# Monitoring
psutil==6.1.0
prometheus-client==0.19.0